requirements.txt
# CMML Market Scan Requirements - With LangGraph

# HTTP Requests
requests>=2.32.0

# Environment Variables
python-dotenv>=1.0.1

# Logging
loguru>=0.7.2

# PDF Processing
PyPDF2>=3.0.1

# YAML Configuration
PyYAML>=6.0.1

# Web Server
flask>=2.3.3
flask-cors>=4.0.0
werkzeug>=2.3.7

# LangGraph and dependencies - latest versions
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
